backend/requirements.txt
openai
langchain-text-splitters
python-dotenv
boto3
pypdf
psycopg2-binary
pydantic
pandas

# main.py
fastapi
uvicorn
jinja2
websockets
docker
